Refactor convert_s and add to avoid Verifier::emitToString (bug 517127 r=tharwood+)
Verifier::emitToString() is a holdover from before the CodeWriter refactoring,
and is only used by the jit.

This patch eliminates it; the type-sniffing optimizations it performed are
pushed down into the jit, and consolidated.

the patch removes the extra cases for OP_esc_xattr, OP_esc_xelem, and
OP_convert_s from writeOp1.

finally, this patch removes the internal opcode OP_concat, which was never
exposed in abc and just gummed up the works.